Transferred previewing to app level
