+++ lib/Resources/globals.dart
@@ -13,7 +13,7 @@ String GFYCAT_GET_URL = "https://api.gfycat.com/v1/gfycats/";
 String GFYCAT_CLIENT_ID = "2__lD9Ci";
 String GFYCAT_CLIENT_SECRET = "waadJXMtWmfHC45OeMvE9lDrKkhQ9XCR0xLMbaFTuINQPjd4s0mcrnnBN8cMmuAr";
 
-String currentSubreddit = "gifs";
+String currentSubreddit = "nsfw_gifs";
 
 @HiveType(adapterName: "PostViewAdapter")
 enum PostView{
