+++ lib/UI/Comments/comment_list.dart
@@ -35,76 +35,10 @@ class CommentList extends StatefulWidget {
 }
 
 
-class CommentListState extends State<CommentList> with SingleTickerProviderStateMixin, PreviewCallback{
+class CommentListState extends State<CommentList> with SingleTickerProviderStateMixin{
   final Submission submission;
 
   CommentListState(this.submission);
-
-  bool isPreviewing = false;
-  var previewUrl = "";
-
-  OverlayState state;
-  OverlayEntry entry;
-
-  @override
-  void initState() {
-    super.initState();
-    state = Overlay.of(context);
-    entry = OverlayEntry(
-        builder: (context) => new GestureDetector(
-              child: new Container(
-                  width: 400.0,
-                  height: 500.0,
-                  child: new Container(
-                      child: Image(
-                          image: AdvancedNetworkImage(
-                        previewUrl,
-                        useDiskCache: true,
-                        cacheRule: CacheRule(maxAge: const Duration(days: 7)),
-                      )),
-                      color: Color.fromARGB(200, 0, 0, 0),
-                  ),),
-              onLongPressUp: () {
-                hideOverlay();
-              },
-            ));
-  }
-
-  showOverlay() {
-    if (!isPreviewing) {
-      state.insert(entry);
-      isPreviewing = true;
-    }
-  }
-
-  hideOverlay() {
-    if (isPreviewing) {
-      entry.remove();
-      state.deactivate();
-      isPreviewing = false;
-    }
-  }
-
-  @override
-  void previewEnd() {
-    if (isPreviewing) {
-      previewUrl = "";
-      // previewController.reverse();
-      hideOverlay();
-    }
-  }
-
-  @override
-  void view(String s) {}
-
-  @override
-  void preview(String url) {
-    if (!isPreviewing) {
-      previewUrl = url;
-      showOverlay();
-      //previewController.forward();
-    }
-  }
 
   CommentsBloc bloc;
 
@@ -132,7 +66,7 @@ class CommentListState extends State<CommentList> with SingleTickerProviderState
                           fontSize: 26.0,
                         ),
                       )
-                      : postInnerWidget(recentlyViewed[i-1], this, PostView.Compact,);
+                      : postInnerWidget(recentlyViewed[i-1], PreviewSource.Comments, PostView.Compact,);
                   },
                 ),
               ),
@@ -144,7 +78,7 @@ class CommentListState extends State<CommentList> with SingleTickerProviderState
                       new SliverToBoxAdapter(
                         child: new Hero(
                           tag: 'post_hero ${submission.id}',
-                          child: new postInnerWidget(submission, this, PostView.ImagePreview),
+                          child: new postInnerWidget(submission, PreviewSource.Comments, PostView.ImagePreview),
                         ),
                       ),
                       
@@ -165,7 +99,6 @@ class CommentListState extends State<CommentList> with SingleTickerProviderState
                       )
                     ],
                   ),
-                  onTapUp: hideOverlay(),
                   onHorizontalDragUpdate: (DragUpdateDetails details) {
                     if (details.delta.direction < 1.0 &&
                         details.delta.dx > 30) {
