+++ lib/UI/interfaces/previewCallback.dart
@@ -3,3 +3,9 @@ abstract class PreviewCallback{
   void view(String u);
   void previewEnd();
 }
+
+// Source view of preview. Needed for deciding whether to show every line of selftext
+enum PreviewSource{
+  Comments,
+  PostsList
+}
